project-2-jwt-hashing-oauth/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
PyJWT==2.8.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
pydantic==2.5.3
python-dotenv==1.0.0
